requirements.txt
dj-database-url==0.4.1
Django==1.11
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.3.0
discord.py==0.16.9
aiohttp
websockets
PyNaCl
youtube_dl